Add ImportSongs class and integrate with CLInterface
Added ImportSongs class to aid in the import and parsing of song data.
This class is static and should not be instantiated. Additionally,
updated javadoc comments and added a testing block within
CLInterface.runSession().

* Added a new ImportSongs class and necessary imports
* Moved REMOVED_WORDS and removedWords() from the Driver class and into
  the ImportSongs class
* Added fetchSongs() method to ImportSongs class for the import of
  external song data files
* Added a testing block within runSession() to ensure song imports are
  importing as expected
* Updated file javadoc to note supplemental authoring within the file
* Updated and clarified several CLInterface javadoc comments, notably
  to clarify that this is a command-line based program

These additions and changes improve the organization and flexibility of
the workflow, and aim to solidify CLInterface.java's purpose as the
primary handler and controller for frontend-to-backend integration.